Update proxyhandlers.py
Allowing exceptions to propagate up by removing the await proc.kill() line. This way, when an error occurs in the Dash app, the exception details will be captured, and JupyterHub or the underlying infrastructure can handle the error. This change should prevent the 500 Internal Server Error and would show detailed logs